# specs/openapi/inbound/payment.yaml
---
openapi: "3.0.3"
info:
  title: "JEMIL Agency API"
  description: "API de gestion des agences de transport partenaires JEMIL.
    
    Ce contrat est la source de vérité entre le backend et les clients
    (app mobile Android, dashboard web agence, équipe QA).
    
    **Règle** : toute modification de ce fichier nécessite une revue.
    Le code est généré depuis ce spec — pas l'inverse."
  version: "2026-W01"
  contact:
    name: "JEMIL Dev Team"
servers:
- url: "https"
- url: "http"
tags:
- name: "Demo"
  description: "Demo operations"
- name: "Agency"
  description: "Agencies operations"
paths:
  /demo:
    get:
      tags:
      - "Demo"
      summary: "Fetch all Demos"
      description: "Fetch all Demos"
      operationId: "fetchAllDemo"
      security:
      - oAuth:
        - "demo:read:all"
      parameters:
      - name: "fieldsToExtractCode"
        in: "query"
        required: true
        schema:
          type: "string"
        description: "specify the fields that will be present on the response"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DemoDTO"
    post:
      tags:
      - "Demo"
      summary: "Create a new Demo"
      description: "Create a new Demo"
      operationId: "createDemo"
      security:
      - oAuth:
        - "demo:create"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDemoDTO"
      responses:
        "201":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponseDTO"
  /demo/{demoId}:
    get:
      tags:
        - "Demo"
      summary: "Fetch demo by id"
      description: "Fetch demo by id"
      operationId: "fetchDemoById"
      security:
        - oAuth:
            - "demo:read"
      parameters:
        - name: "demoId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
          description: "represents the id of demo to fetch."
        - name: "fieldsToExtractCode"
          in: "query"
          required: true
          schema:
            type: "string"
          description: "specify the fields that will be present on the response"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoDTO"
        "404":
          description: "SPRING_BOOT_TEMPLATE-404-001"
  /agency:
    post:
      tags:
        - "Agency"
      summary: "Enregistrer une nouvelle agence"
      description: "|
        Enregistre une agence de transport comme partenaire JEMIL.
        L'agence est créée avec le statut ACTIVE par défaut."
      operationId: "registerAgency"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgencyDTO'
      responses:
        '201':
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDTO'
        '400':
          description: Données invalides ou règle métier violée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '409':
          description: "Une agence avec ce nom existe déjà dans cette ville"
    get:
      tags:
        - "Agency"
      summary: "Lister les agences actives"
      description: "Retourne toutes les agences actives, filtrables par ville."
      operationId: "getAllAgencies"
      parameters:
        - name: "city"
          in: query
          required: false
          description: "Filtrer par ville (insensible à la casse)"
          schema:
            type: "string"
            example: "Douala"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgencyDTO'

  /agency/{agencyId}:
    get:
      tags:
        - "Agency"
      summary: "Récupérer une agence par son ID"
      operationId: "getAgencyById"
      parameters:
        - name: agencyId
          in: path
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDTO'
        '404':
          description: Agence non trouvée

  /agency/{agencyId}/routes:
    post:
      tags:
        - "Agency"
      summary: "Ajouter une route à une agence"
      operationId: "addRoute"
      parameters:
        - name: "agencyId"
          in: path
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRouteDTO'
            example:
              origin: "Douala"
              destination: "Yaoundé"
              totalSeats: 50
      responses:
        '201':
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDTO'
        '400':
          description: "Données invalides"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: "Agence non trouvée"
components:
#  securitySchemes:
#    bearerAuth:
#      type: "http"
#      scheme: "bearer"
#      bearerFormat: "JWT"
#    oAuth:
#      type: "oauth2"
#      description: "This API uses OAuth 2 with the Authorization Code grant flow."
#      flows:
#        authorizationCode:
#          authorizationUrl: "https://auth.oauth.lao-sarl.cm"
#          tokenUrl: "https://token.oauth.lao-sarl.cm"
#          scopes:
#            demo:read:all: "Needed to fetch all Demo objects"
#            demo:create: "Needed when to create a new Demo object"
  schemas:
    DemoDTO:
      type: "object"
      required:
      - "id"
      - "name"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
    CreateDemoDTO:
      type: "object"
      required:
      - "name"
      properties:
        name:
          type: "string"
    CreationResponseDTO:
      type: "object"
      required:
      - "newId"
      properties:
        newId:
          type: "string"
          format: "uuid"
    PhoneNumberDTO:
      type: "object"
      properties:
        countryCode:
          type: "string"
          description: "the international code for the country"
          example: "237 for Cameroon"
        number:
          type: "string"
          description: "the phone number"
          example: 653492410
    AddressDTO:
      type: "object"
      properties:
        city:
          type: "string"
          description: "the city of collector"
        district:
          type: "string"
          description: "the district of collector"
    RegisterAgencyDTO:
      type: "object"
      required:
        - "name"
        - "city"
      properties:
        name:
          type: "string"
          minLength: "2"
          maxLength: "100"
          description: "Nom commercial de l'agence"
          example: "Global Voyages"
        address:
          $ref: "#/components/schemas/AddressDTO"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumberDTO"

    AddRouteDTO:
      type: "object"
      required:
      - "origin"
      - "destination"
      - "totalSeats"
      properties:
        origin:
          type: "string"
          description: "Ville de départ"
          example: "Douala"
        destination:
          type: "string"
          description: "Ville d'arrivée"
          example: "Yaoundé"
        totalSeats:
          type: "integer"
          minimum: "1"
          maximum: "100"
          description: "Nombre total de places dans le véhicule"
    AgencyDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Identifiant unique de l'agence"
        name:
          type: "string"
        address:
          $ref: "#/components/schemas/AddressDTO"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumberDTO"
        status:
          $ref: "#/components/schemas/StatusDTO"
        routeCount:
          type: "integer"
    StatusDTO:
      enum:
        - "ACTIVE"
        - "SUSPENDED"
        - "INACTIVE"
    RouteDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        agencyId:
          type: "string"
          format: "uuid"
        origin:
          type: "string"
        destination:
          type: "string"
        totalSeats:
          type: "integer"
        active:
          type: "boolean"
    ErrorDTO:
      type: "object"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
        status:
          type: "integer"
        error:
          type: "string"
          description: "Code d'erreur technique"
          example: "DOMAIN_RULE_VIOLATION"
        message:
          type: "string"
          description: "Message lisible par un humain"
          example: "Le nom de l'agence est obligatoire."